web/requirements.txt
Pillow
aiohttp